code refactor and test setup
