+++ package.json
@@ -26,16 +26,27 @@
     "babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
     "babel-plugin-transform-es3-property-literals": "^6.22.0",
     "babel-preset-env": "^1.6.1",
+    "chai": "^4.1.2",
     "gulp": "^3.9.1",
     "gulp-clean": "^0.4.0",
     "gulp-connect": "^5.2.0",
     "gulp-eslint": "^4.0.2",
     "gulp-header": "^2.0.1",
     "gulp-inject": "^4.3.0",
-    "gulp-replace": "^0.6.1",
     "gulp-rename": "^1.2.2",
+    "gulp-replace": "^0.6.1",
     "gulp-uglify": "^3.0.0",
+    "karma": "^3.0.0",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-mocha": "^1.3.0",
+    "karma-requirejs": "^1.1.0",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^3.0.5",
     "lodash": "^4.17.5",
+    "mocha": "^5.2.0",
+    "requirejs": "^2.3.6",
+    "sinon": "^6.3.4",
     "webpack": "^3.0.0",
     "webpack-stream": "^4.0.0",
     "yargs": "^11.0.0"
