+++ test/spec/environment_spec.js
@@ -1,0 +1,12 @@
+import chai from "chai";
+
+const assert = chai.assert;
+
+describe('Array', function() {
+  describe('#indexOf()', function() {
+    it('should return -1 when the value is not present', function() {
+      assert.equal([1,2,3].indexOf(4), -1);
+    });
+  });
+});
+
