+++ test/test_index.js
@@ -1,0 +1,2 @@
+var testsContext = require.context('.', true, /_spec$/);
+testsContext.keys().forEach(testsContext);
