+++ webpack.conf.js
@@ -3,7 +3,7 @@ var path = require('path');
 var webpack = require('webpack');
 
 module.exports = {
-  devtool: 'source-map',
+  devtool: 'inline-source-map',
   resolve: {
     modules: [
       path.resolve('.'),
@@ -11,7 +11,7 @@ module.exports = {
     ],
   },
   output: {
-    filename: "creative.js"
+    filename: 'creative.js',
   },
   module: {
     rules: [
